% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventtimeresid.R
\name{etime.resid}
\alias{etime.resid}
\title{Cox-Snell residuals}
\usage{
etime.resid(
  fit,
  plot = FALSE,
  conf.int = 0.95,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL
)
}
\arguments{
\item{fit}{\code{etime.reg} fit.}

\item{plot}{returns residual plot if \code{TRUE} (default is \code{FALSE}).}

\item{conf.int}{confidence level for the pointwise confidence intervals of the residuals (if \code{plot = TRUE}).}

\item{xlim}{x limits (a vector of length 2 of the form \code{c(minimum,maximum)}) of the plot (optional).}

\item{ylim}{y limits (a vector of length 2 of the form \code{c(minimum,maximum)}) of the plot (optional).}

\item{main}{main title (character expression) of the plot (optional).}
}
\value{
\code{residuals} Cox-Snell residuals.

\code{plot} residual plot if \code{plot=TRUE}.
}
\description{
Cox-Snell residuals of the \code{etime.reg} fit.
}
\details{
If \code{plot = TRUE}, this function returns both residuals and a residual plot. 
    The plot shows the unit slop line (in grey), residuals (solid circles) and pointwise confidence intervals (dashed lines).
    The plot of the Cox-Snell (hazard-based) residuals should be roughly a straight line 
    with unit slope when the model is adequate. 
    See Section 6.2 of Lawless and Section 3.7.3 of Cook and Lawless for details.
}
\examples{
# Example 1: Recurrent Event Analysis
  library(frailtypack)
  data(readmission)
  fit.gg <- etime.reg(c(t.start, t.stop, event) ~ sex + chemo + charlson,
      surv.model = "ggamma", Data = readmission)
  fit.wph <- etime.reg(c(t.start, t.stop, event) ~ sex + chemo + charlson,
      surv.model = "weibullph", Data = readmission)
  etime.resid(fit.wph, plot = TRUE)
  par(mfrow = c(1, 2))
  etime.resid(fit.wph, plot = TRUE, xlim = c(0, 5.25),
    ylim = c(0, 5.25),  main = "Weibull PH")
  etime.resid(fit.gg, plot = TRUE, xlim = c(0, 5.25),
     ylim = c(0, 5.25),  main = "Generalized Gamma AFT")

# Example 2: Non-recurrent Data (Right Censored)
  library(survival)
  fit.ln <- etime.reg(c(time, status) ~ karno + diagtime + age + prior + celltype + trt,
      Data = veteran, surv.model = "lnormal")
  etime.resid(fit.ln, plot = TRUE, main = "Log-normal AFT")
}
\references{
Cook RJ and Lawless J, The statistical analysis of recurrent events, Springer, 2007.

Lawless J, Statistical models and methods for lifetime data, Wiley, 2003.
}
\seealso{
\code{\link{etime.reg}}, \code{\link{LR.test}}.
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Cox-Snell}
\keyword{residual}
