% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMcriteria.R
\name{jm.icriteria}
\alias{jm.icriteria}
\title{Information criteria (DIC and WAIC)}
\usage{
jm.icriteria(jmfit, posterior.mean = TRUE)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{posterior.mean}{if \code{TRUE} (default), the posterior mean is used to compute
DIC, otherwise posterior median is used. Note that this argument
is used only for DIC calculation.}
}
\value{
\code{DIC:} the values of log predictive density (lpd) or log-likelihood given a point estimate of the 
        fitted model, effective number of parameters (p), and DIC. If \code{posterior.mean = TRUE},
        the posterior mean is used as a point estimate, otherwise posterior median is used.

\code{WAIC:} the values of log pointwise predictive density (lppd) evaluated using the posterior 
        simulations, effective number of parameters (p), and WAIC.
}
\description{
Computes DIC and WAIC for a Bayesian joint model fit. It uses a \code{\link{jm.reg}} fit as its argument.
}
\details{
This function uses the posterior draws from class "stanfit". See Gelman et al. (2014) 
      for a description of DIC and WAIC from a Bayesian perspective.
}
\examples{
  # Example: pbc data
  lme.fit <- lme(log(bilirubin) ~ drug + ns(futime, 2),
    data = pbc.long, random = ~ ns(futime, 2) | id)
  surv.fit <- coxph(Surv(st, status2) ~ drug * age, data = pbc.surv, x = TRUE)
  jmfit.ll2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "llogistic", timevar = "futime", form = "riz")
  jm.icriteria(jmfit.ll2)
  jmfit.wph2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "futime", form = "riz")
  jm.icriteria(jmfit.wph2)
}
\references{
Gelman A, Hwang J, and Vehtari A, Understanding predictive information criteria 
    for Bayesian models, Statistics and Computing, 24: 997-1016, 2014.
}
\seealso{
\code{\link{jm.plots}}, \code{\link{jm.reg}}, \code{\link{jm.resid}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{DIC,}
\keyword{Joint}
\keyword{MCMC,}
\keyword{WAIC}
\keyword{modeling,}
