% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmpredict.R
\name{jm.lppredict}
\alias{jm.lppredict}
\title{Marginal predictions for the longitudinal outcome of a joint model}
\usage{
jm.lppredict(jmfit, newdata)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{newdata}{a data frame in which to look for variables with which to predict.
The data frame must have columns that correspond by name to the
original regressors used in the longitudinal part of the joint model fit.}
}
\value{
This function returns posterior summaries of X(s)\eqn{\alpha} (marginal predictions).
}
\description{
Given a \code{\link{jm.reg}} fit and a data frame, this function
     produces marginal predictions (population-level) for the longitudinal outcome.
}
\details{
Posterior summaries of X(s)\eqn{\alpha} (marginal predictions)
    are obtained from the MCMC samples, where X(s) is the fixed-effects
    design matrix and \eqn{\alpha} is the fixed-effects
    coefficient vector for the longitudinal model.
}
\examples{
  # Example: pbc data
  lme.fit <- lme(log(bilirubin) ~ drug + ns(futime, 2),
    data = pbc.long, random = ~ ns(futime, 2) | id)
  surv.fit <- coxph(Surv(st, status2) ~ drug * age, data = pbc.surv, x = TRUE)
  jmfit.gg2 <-  jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "ggamma", timevar = "futime", form = "riz")
  jm.summary(jmfit.gg2)
  newdata <-  expand.grid(futime = seq(min(pbc.long$futime), max(pbc.long$futime), 
      len = 25), drug = 0:1)
  pred <- jm.lppredict(jmfit = jmfit.gg2, newdata = newdata)
  pred
  # Plot of the predicted values
  futime <- pred[[1]][, 1]
  post.mean <- pred[[1]][, 2]
  lower <- pred[[1]][, 4]
  upper <- pred[[1]][, 8]
  drug <- factor(newdata$drug, levels = c(0, 1), labels = c("placebo", "D-penicillamine"))
  # Use the xyplot function from package 'lattice'
  library(lattice)
  xyplot(post.mean + lower + upper ~ futime | drug,
    type = "l", col = rep(1, 3), lty = c(1, 3, 3),
    lwd = 2,  ylab = "log(bilirubin)",  xlab = "futime")
}
\seealso{
\code{\link{jm.lspredict}}, \code{\link{jm.surv}}, \code{\link{jm.reg}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{Longitudinal}
\keyword{MCMC,}
\keyword{Prediction,}
\keyword{modeling,}
