% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmpredict.R
\name{jm.lspredict}
\alias{jm.lspredict}
\title{Subject-specific predictions for the longitudinal outcome of a joint model}
\usage{
jm.lspredict(
  jmfit,
  newdata,
  n.sample = NULL,
  ptime = NULL,
  plot = TRUE,
  posterior.mean = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  adapt_delta = 0.9,
  max_treedepth = 15,
  warmup = 200,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{newdata}{a data frame in which to look for variables with which to predict.
The data frame must have columns that correspond by name to the
original variables used to fit the joint model.}

\item{n.sample}{an integer denoting how many MCMC samples to use. Default is \code{min(200, n.iter)}, where
n.iter is the number of MCMC iterations (for all chains combined).}

\item{ptime}{a numeric vector of times at which predictions 
are to be computed.
If \code{NULL}, \code{ptime = seq(ptime0, floor(max.st), length=25)} is used,
where \code{ptime0} is the last time the subject provided a longitudinal measurement, and
\code{max.st} is \code{floor(max(event times))}.}

\item{plot}{if \code{TRUE}, a plot of the predictions is produced.}

\item{posterior.mean}{if \code{TRUE}, posterior means of the predictions are used in the plot,
otherwise posterior medians are used.}

\item{xlim}{x limits (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).}

\item{ylim}{y limits (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).}

\item{xlab}{a label (character expression) for the x axis (optional).}

\item{ylab}{a label (character expression) for the y axis (optional).}

\item{main}{main title (character expression) of the plot (optional).}

\item{adapt_delta}{the target average proposal acceptance probability during Stan’s adaptation period.}

\item{max_treedepth}{a positive integer specifying the maximum treedepth.}

\item{warmup}{a positive integer specifying the number of burnin iterations.}

\item{seed}{the seed for random number generation.}
}
\value{
Posterior summaries of the subject-specific predictions.
        A plot of the predictions is produced if \code{plot = TRUE}, with
        the predictions displayed by the solid curve, the pointwise Bayesian intervals 
        displayed by the dashed curves, and \code{ptime[1]} displayed by the dashed vertical line.
        Note that with the default definition of \code{ptime}, \code{ptime[1]} is the last time the subject 
        provided a longitudinal measurement.
}
\description{
Given a \code{\link{jm.reg}} fit and a data frame, this function
     produces subject-specific predictions for the longitudinal outcome.
}
\details{
This function computes subject-specific predictions for the longitudinal
      outcome based on the joint model. This is accomplished with a Monte Carlo simulation scheme,
      similar to the one described in \code{\link{jm.surv}}. Let \eqn{\theta^{(m)}} be a 
      realization taken randomly from the posterior simulations of the joint model parameters
      \eqn{(m = 1, 2, ...,} \code{n.sample)}. Also, let
      \eqn{b^{(m)}} be a realization of the random-effects vector b from its 
      posterior distribution given \eqn{\theta^{(m)}} (see the MCMC scheme described in \code{\link{jm.surv}}). 
      Posterior summaries of \eqn{x'_i(s)\alpha + w'_i(s)b_i} (subject-specific prediction)
      at time \eqn{s} are derived from the MCMC sample \eqn{\{(\theta^{(m)}, b^{(m)}),m = 1, 2, ...,} \code{n.sample\}}.
}
\examples{
  # Example: pbc data
  lme.fit <- lme(log(bilirubin) ~ drug + ns(futime, 2), data = pbc.long,
    random = ~ ns(futime, 2) | id)
  surv.fit <- coxph(Surv(st, status2) ~ drug * age, data = pbc.surv, x = TRUE)
  jmfit.wph2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "futime", form = "riz")
  jm.summary(jmfit.wph2)
  newdata2 <- data.frame(pbc.long[pbc.long$id == 2, ])
  newdata6 <- data.frame(pbc.long[pbc.long$id == 6, ])
  par(mfrow = c(1, 2))
  pred2 <- jm.lspredict(jmfit = jmfit.wph2, newdata = newdata2)
  pred6 <- jm.lspredict(jmfit = jmfit.wph2, newdata = newdata6)
  # For comparison, use the same xlim and ylim in the two plots.
  par(mfrow = c(1, 2))
  pred2 <-  jm.lspredict(jmfit = jmfit.wph2, newdata = newdata2,
      xlim = c(0, 14),  ylim = c(-2.25, 4))
  pred6 <- jm.lspredict(jmfit = jmfit.wph2, newdata = newdata6,
      xlim = c(0, 14), ylim = c(-2.25, 4))
}
\references{
Rizopoulos D, The R package JMbayes for fitting joint models for longitudinal and time-to-event data Using MCMC, 
          Journal of Statistical Software, 72(7): 1-45, 2016.
}
\seealso{
\code{\link{jm.lppredict}}, \code{\link{jm.reg}}, \code{\link{jm.surv}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{Longitudinal}
\keyword{MCMC,}
\keyword{Prediction,}
\keyword{modeling,}
