% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMreffects.R
\name{jm.reffects}
\alias{jm.reffects}
\title{Posterior summaries of the random effects from a joint model fit}
\usage{
jm.reffects(jmfit, rhat.plot = FALSE, neff.plot = FALSE)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{rhat.plot}{if \code{TRUE}, produce a plot for rhats.}

\item{neff.plot}{if \code{TRUE}, produce a plot for ratios of effective sample size to total sample size.}
}
\value{
\code{b:} posterior summaries of the random effects.

\code{plots:} plots if \code{rhat.plot = TRUE} and/or \code{neff.plot = TRUE}.
}
\description{
Returns posterior summary of the random effects. Note that the random effects are shared between the 
    longitudinal and the survival components, and the link between these two processes via the random effects 
    is commonly known as \emph{latent association}. The formulation of the joint model
    is described in \code{\link{jm.reg}}.
}
\details{
The random effects are monitored in MCMC simulations in Stan. The summary method in rstan is used
       to produce the posterior summaries of the random effects.
}
\examples{
  # Example: pbc data
  lme.fit <- lme(log(bilirubin) ~ drug + ns(futime, 2),
    data = pbc.long, random = ~ ns(futime, 2) | id)
  surv.fit <- coxph(Surv(st, status2) ~ drug * age, data = pbc.surv, x = TRUE)
  # Bayesian estimation
  jmfit.wph2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "futime", form = "riz")
  jm.reffects(jmfit.wph2, rhat.plot = TRUE, neff.plot = TRUE)
}
\seealso{
\code{\link{jm.reg}}, \code{\link{jm.summary}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{MCMC,}
\keyword{Posterior}
\keyword{Random}
\keyword{effects}
\keyword{modeling,}
\keyword{summary,}
