% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMresid.R
\name{jm.resid}
\alias{jm.resid}
\title{Cox-Snell residuals for a jm.reg fit}
\usage{
jm.resid(
  jmfit,
  posterior.mean = TRUE,
  n.sample = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{posterior.mean}{returns posterior means of the residuals if \code{posterior.mean = TRUE}, otherwise
returns posterior medians of the residuals.}

\item{n.sample}{an integer denoting how many MCMC samples to use. Default is \code{min(200, n.iter)}, where
n.iter is the number of MCMC iterations (for all chains combined).}

\item{xlim}{x limits (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).}

\item{ylim}{y limits (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).}

\item{xlab}{a label (character expression) for the x axis (optional).}

\item{ylab}{a label (character expression) for the y axis (optional).}

\item{main}{main title (character expression) of the plot (optional).}

\item{seed}{the seed for random number generation to draw \code{n.sample} samples.
Default is \code{sample.int(.Machine$integer.max, 1)}.}
}
\value{
\code{Cox-Snell Residuals:} posterior expectation of the residuals and \eqn{-\textrm{log}\textrm{S}_{\textrm{KM}}}(rediduals).

\code{plot:} see Details.
}
\description{
Returns posterior summaries of the Cox-Snell residuals for a \code{\link{jm.reg}} fit.
  It also produces a residual plot.
}
\details{
A random sample of size \code{n.sample} is drawn from the posterior simulations, and the
  Cox-Snell residuals are computed for each sample. Following Rizopoulos and Ghosh (2011), the estimate of 
  the posterior expectation is obtained as the MCMC sample mean (or median) of the residuals 
  at each observed time point.
  A plot of residual vs. \eqn{-\textrm{log}\textrm{S}_{\textrm{KM}}}(redidual) is then produced (solid circles), 
  where \eqn{\textrm{S}_{\textrm{KM}}}(redidual) is the Kaplan-Meier estimate of the survivor function.
 Uncertainty in the plot can be assessed through the \code{n.sample} sets of residuals obtained
 from the posterior simulations (Zhou and Hanson, 2018). These are shown in the plot as grey lines.
 The plot also shows the unit slop line (in black). The plot of the Cox-Snell residuals should be roughly a straight line 
    with unit slope when the survival sub-model is adequate.
}
\examples{
  # Example: AIDS data from package 'JM'
  library(JM)
  data(aids.id)
  data(aids)
  surv.fit <- coxph(Surv(Time, death) ~ drug + gender + prevOI + AZT,
          data = aids.id, x = TRUE)
  lme.fit <- lme(CD4 ~ obstime + obstime:drug + gender + prevOI + AZT,
      random =  ~ obstime | patient, data = aids)
  # Weibull PH with form = "riz"
  jmfit.wph0 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "obstime", form = "riz")
  jm.resid(jmfit.wph0)
}
\references{
Rizopoulos D and Ghosh P, A Bayesian semiparametric multivariate joint model for multiple 
      longitudinal outcomes and a time‐to‐event, Statistics in Medicine, 30: 1366-1380, 2011.

Zhou H and Hanson T, A unified framework for fitting Bayesian
      semiparametric models to arbitrarily censored survival data, including spatially referenced data, 
      Journal of the American Statistical Association, 113(522), 571-581, 2018.
}
\seealso{
\code{\link{jm.icriteria}}, \code{\link{jm.plots}}, \code{\link{jm.reg}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Cox-Snell}
\keyword{Joint}
\keyword{MCMC}
\keyword{modeling,}
\keyword{residual,}
