% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmsim.R
\name{jm.sim}
\alias{jm.sim}
\title{Simulate from joint models}
\usage{
jm.sim(
  surv.formula,
  lme.formula,
  surv.model,
  form,
  surv.par,
  lme.par,
  times,
  timevar,
  Data
)
}
\arguments{
\item{surv.formula}{a one-sided formula of the form \code{~ z1 + z2 + ... + zp}, where 
\code{z1, z2, ..., zp} are baseline covariates for the time-to-event process.}

\item{lme.formula}{a one-sided formula of the form \code{~ times} or \code{~ times + x2 + x3 + ... + xq}
for the fixed-effects component of the longitudinal model, where 
\code{times} is the time variable (time points at which longitudinal measurements are taken) and
\code{x2, x3, ..., xq} are time-independent covariates.}

\item{surv.model}{the survival model to be used to describe the event process. Available options are given as follow.
 \itemize{ 
    \item AFT Models: \code{"weibull"}, \code{"lnormal"}, \code{"llogistic"}, \code{"eweibull"} and \code{"ggamma"}
          for Weibull, log-normal, log-logistic, exponentiated Weibull and generalized gamma distributions, respectively. 
    \item PH Models: \code{"weibullph"} and \code{"gllph"} for Weibull and generalized log-logistic distributions, respectively.
  }
The parameterization of these distributions is described in \code{\link{etime.reg}}.}

\item{form}{a character string to describe the formulation of the joint model. Available options are
\code{form = "rizopoulos"} (can be abbreviated as "riz") for the formulation described by Rizopoulos (2012), and
\code{form = "henderson"} (can be abbreviated as "hen") for the formulation proposed by Henderson et al. (2000). 
See \code{\link{jm.reg}} for details.}

\item{surv.par}{a named list of the form \code{list(beta = a vector, phi = a scalar, logrho = a scalar,} 
\code{logkappa = a scalar, loggamma = a scalar, status.rho = a scalar)} for the time-to-event model parameters, 
where \code{beta} is the vector of regression coefficients for the baseline covariates as specified in
\code{surv.formula} (without an intercept), \code{phi} is the association parameter, 
\code{logrho} = log\eqn{\rho}, \code{logkappa} = log\eqn{\kappa}, \code{loggamma} = log\eqn{\gamma},
and \code{status.rho} is the rate
parameter of the exponential distribution from which censored data are generated.
See \code{\link{etime.reg}} for parameterizations of the time-to-event distributions
in terms of \eqn{\rho}, \eqn{\kappa} and \eqn{\gamma}.}

\item{lme.par}{a named list of the form \code{list(alpha = a vector, sigma = a scalar, rand_cov = a 2 x 2 covariance matrix)} 
for the longitudinal model parameters, 
where \code{alpha} is the vector of fixed-effects regression coefficients as specified in
\code{lme.formula} (includes an intercept), \code{sigma} is the 
standard deviation of measurement errors (i.e., the value of \eqn{\sigma} for \eqn{\epsilon} ~ \eqn{N(0,\sigma^2)}),
and \code{rand_cov} is the covariance matrix of the random effects \eqn{b_{0}} and \eqn{b_{1}}. 
Note that only a simple linear model \eqn{b_{0} + b_{1}} \code{times} is allowed for random effects.}

\item{times}{a numeric vector for the time points at which longitudinal measurements are planned to be taken.}

\item{timevar}{the time variable (a character string) of the longitudinal model. For example,
if \code{lme.formula = ~ times + x1}, then \code{timevar = "times"}.}

\item{Data}{a data frame containing the baseline covariates \code{z}'s for the time-to-event model and
the time-independent covariates \code{x}'s for the longitudinal model. For \eqn{n} subjects, it must
must have \eqn{n} rows.}
}
\value{
A list with components:

\code{b:} random effects (a matrix).

\code{long.data:} simulated longitudinal data.

\code{surv.data:} simulated event time data.
}
\description{
This function simulates longitudinal responses and event times from joint models.
    Available options for the time-to-event submodel are Weibull AFT, log-logistic AFT, log-normal AFT, 
    exponentiated Weibull AFT, generalized gamma AFT, Weibull PH and generalized log-logistic PH.
    For the longitudinal process, the linear mixed-effects model is assumed, 
    with \emph{a simple linear model for the random-effects component}
    (i.e., \eqn{b_0 + b_1} \code{times}, where \eqn{b_0} and \eqn{b_1} are the random intercept and random slope,
    respectively).
}
\details{
This function simulates longitudinal responses and event times from joint models.
    The longitudinal model is of the form \eqn{y = x'(s)\alpha +w'(s)b + \epsilon},
    where \eqn{x'(s)\alpha=\alpha_0 + \alpha_1 s} or
    \eqn{\alpha_0 + \alpha_1 s + \alpha_2 x_2 + \alpha_3 x_3 + ... + \alpha_q x_q},
    \eqn{w'(s)b = b_0 + b_1 s}, \eqn{\epsilon} ~ \eqn{N(0,\sigma^2)} and \eqn{b} ~ \eqn{N(0, \Sigma_b)}.
    For an AFT event process, the survivor function is of the form
    \eqn{S_0(\int_0^t} exp\eqn{[-(z'\beta+\phi(x'(u)\alpha +w'(u)b)]du)} for \code{form = "rizopoulos"}, 
    and \eqn{S_0(\int_0^t} exp\eqn{[-(z'\beta+\phi w'(u)b)]du)} for \code{form = "henderson"}.
    For a PH event process, the survivor function is of the form exp[\eqn{-\Lambda(t)}], where
    \eqn{\Lambda(t)=\int_0^t \lambda_0(u)} exp\eqn{[z'\beta+\phi(x'(u)\alpha +w'(u)b)]du} for \code{form = "rizopoulos"},
    and \eqn{\Lambda(t)=\int_0^t \lambda_0(u)} exp\eqn{[z'\beta+\phi w'(u)b]du} for \code{form = "henderson"}.
    Event times are drawn by solving the equation \code{survivor function = U} for \eqn{t}, where
    \code{U} is a Uniform(0, 1) variable. Note that
    the amount of censoring can be controlled by changing the value of \code{status.rho}.
}
\examples{
  n <- 100
  surv.formula <- ~ z1 + z2
  lme.formula <- ~ times + x1 + x2
  cor.mat <- matrix(c(1, -0.2, -0.2, 1), ncol = 2, byrow = TRUE)
  sd <- c(0.5, 0.1)
  rand_cov <- (sd \%*\% t(sd)) * cor.mat
  lme.par <- list(alpha = c(10, -1, -1, 0.25), sigma = 0.5, rand_cov = rand_cov)
  surv.par <- list(beta = c(-0.5, 0.5), phi = -0.25, logrho = log(0.01),
    logkappa = log(2), status.rho = 0.003)
  z1 <- rbinom(n, 1, 0.5)
  z2 <- rnorm(n)
  # Data with all covariate information
  Data <- data.frame(z1 = z1, z2 = z2, x1 = z1, x2 = z2)
  # Time points at which longitudinal measurements are planned to be taken
  times <- c(0, 1, 2, 3, 4)
  timevar <- "times"
  jmsim.dat <- jm.sim(surv.formula = surv.formula, lme.formula = lme.formula,
    surv.model = "weibullph", form = "hen", surv.par = surv.par,
    lme.par = lme.par, times = times, timevar = timevar, Data = Data)
  surv.data <- jmsim.dat$surv.data
  long.data <- jmsim.dat$long.data
  surv.fit <- coxph(Surv(st, status) ~ z1 + z2, data = surv.data, x = TRUE)
  lme.fit <- lme(y ~ times + x1 + x2, random =  ~ times | id, data = long.data)
  swph <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "times", form = "hen")
  jm.summary(swph)
}
\references{
Henderson R, Diggle P, and Dobson A, Joint modelling of longitudinal measurements and event time data, 
          Biostatistics, 1: 465-480, 2000.

Rizopoulos D, Joint Models for Longitudinal and Time-to-Event Data: 
         With Applications in R, Chapman and Hall/CRC, 2012.
}
\seealso{
\code{\link{jm.reg}}, \code{\link{jm.summary}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{Simulation}
\keyword{model,}
