% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMsummary.R
\name{jm.summary}
\alias{jm.summary}
\title{Summary of a jm.reg fit}
\usage{
jm.summary(jmfit)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}
}
\value{
A summary of the survival data, longitudinal data, the joint model, number of MCMC chains,
   number of iterations per chain after warmup, and computaional time. It also gives

\code{Posterior summary (survival sub-model):} posterior summary for the survival sub-model.

\code{Posterior summary (exp(coef) of the survival sub-model):} exp(coef) for the survival sub-model (for PH models only).

\code{Posterior summary (longitudinal sub-model):} posterior summary for the longitudinal sub-model.
}
\description{
Produces a summary of the Bayesian joint model fit. It uses a \code{\link{jm.reg}} fit as its argument.
}
\details{
This function uses the posterior draws and 
    the posterior summary from class "stanfit".
}
\examples{
  # Example: pbc data
  lme.fit <- lme(log(bilirubin) ~ drug + ns(futime, 2),
    data = pbc.long, random = ~ ns(futime, 2) | id)
  surv.fit <- coxph(Surv(st, status2) ~ drug * age, data = pbc.surv, x = TRUE)
  jmfit.ll2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "llogistic", timevar = "futime", form = "riz")
  jm.summary(jmfit.ll2)
  jmfit.wph2 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "weibullph", timevar = "futime", form = "riz")
  jm.summary(jmfit.wph2)
}
\seealso{
\code{\link{jm.icriteria}}, \code{\link{jm.plots}}, \code{\link{jm.reg}},
         \code{\link{jm.resid}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{MCMC}
\keyword{modeling,}
