% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmpredict.R
\name{jm.surv}
\alias{jm.surv}
\title{Prediction of survival probabilities from a joint model fit}
\usage{
jm.surv(
  jmfit,
  newdata,
  n.sample = NULL,
  st = NULL,
  plot = TRUE,
  posterior.mean = TRUE,
  include.y = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab.left = NULL,
  ylab.right = NULL,
  main = NULL,
  adapt_delta = 0.9,
  max_treedepth = 15,
  warmup = 200,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{jmfit}{a \code{\link{jm.reg}} fit with \code{method = "MCMC"}.}

\item{newdata}{a data frame in which to look for variables with which to predict.
The data frame must have columns that correspond by name to the
original variables used to fit the joint model.}

\item{n.sample}{an integer denoting how many MCMC samples to use. Default is \code{min(200, n.iter)}, where
n.iter is the number of MCMC iterations (for all chains combined).}

\item{st}{a numeric vector of times at which survival probabilities 
P(\eqn{T > t | T > s}, history of the longitudinal response, covariate information) 
are to be computed (see Details).
If \code{NULL}, \eqn{s} is taken to be the last time the 
subject provided a longitudinal measurement, and \code{t = seq(s, max.st, length = 15)}, where
\code{max.st} is \code{floor(max(event times))}. A warning will be given If \eqn{s >} 
the last time the individual provided a longitudinal measurement.}

\item{plot}{if \code{TRUE}, a plot of the survival curve is produced.}

\item{posterior.mean}{if \code{TRUE}, posterior means of the survival probabilities
are used to produce the survival curve, otherwise posterior medians are used.}

\item{include.y}{if \code{TRUE}, the fitted longitudinal profile is included 
in the survival curve plot.}

\item{xlim}{x limits (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).}

\item{ylim}{limits for the left y axis (a vector of length 2 of the form \code{c(minimum, maximum)}) of the plot (optional).
If \code{include.y = TRUE}, the left y axis is for the the fitted longitudinal profile, otherwise
it is for predictions of survival probabilites.}

\item{xlab}{a label (character expression) for the x axis (optional).}

\item{ylab.left}{a label (character expression) for the left y axis (optional). If \code{include.y = TRUE}, 
the left y axis is for the the fitted longitudinal profile, otherwise
it is for predictions of survival probabilites.}

\item{ylab.right}{a label (character expression) for the right y axis (optional). If \code{include.y = TRUE},
the right y axis is for predictions of survival probabilites, otherwise
it is ignored.}

\item{main}{main title (character expression) of the plot (optional).}

\item{adapt_delta}{the target average proposal acceptance probability during Stan’s adaptation period
(see Details).}

\item{max_treedepth}{a positive integer specifying the maximum treedepth (see the Stan manual).}

\item{warmup}{a positive integer specifying the number of burnin iterations (see Details).}

\item{seed}{the seed for random number generation.}
}
\value{
Posterior summaries of the predictions of survival probabilites are returned.
        A plot of the survival curve is produced if \code{plot = TRUE}, with
        the predictions displayed by the solid line and the pointwise Bayesian intervals displayed by the dashed lines.
}
\description{
Dynamic predictions of survival probabilites as described by Rizopoulos (2016, 2020).
}
\details{
The algorithm proposed by Rizopoulos (2020) is used for dynamic predictions.
       The history of the longitudinal response up to time \eqn{s} along with covariate 
       information are taken into consideration for predictions.
       The algorithm is described as follows (see also Rizopoulos (2016, 2020)).
    \enumerate{ 
    \item Randomly take a realization from the posterior simulations of the joint model parameters. 
          Denote it by \eqn{\theta^{(1)}}. 
    \item Given \eqn{\theta^{(1)}}, the history of the longitudinal response, covariate information, 
          and survival up to time \eqn{s}, draw a realization of the random-effects vector \eqn{b} from its posterior distribution.
          For this step, the MCMC algorithm (implemented in Stan) is used with \eqn{n=1}, 
          \code{event time} = \eqn{s} and \code{status} = 0. The number of burnin iterations, adapt_delta
          and max_treedepth in Stan can be controlled using the arguments \code{warmup} (default is 200),
          \code{adapt_delta} (default is 0.90) and \code{max_treedepth} (default is 15).  
          Note that Rizopoulos (2020) used the Metropolis-Hastings algorithm with independent proposals (a 
          properly centered and scaled multivariate t distribution) to estimate \eqn{b}.
    \item Using \eqn{\theta^{(1)}} and \eqn{b}, compute P(\eqn{T > t | T > s}) = P(\eqn{T > t})/P(\eqn{T > s}).
    \item Repeat steps 1-3 \code{n.sample} times.   
         }
    Posterior summeries of the conditional probabilities are derived from the MCMC sample drawn 
    using the above algorithm.
}
\examples{
  # Example: AIDS data from package 'JM'
  library(JM)
  data(aids.id)
  data(aids)
  surv.fit <- coxph(Surv(Time, death) ~ drug + gender + prevOI + AZT,
          data = aids.id, x = TRUE)
  lme.fit <- lme(CD4 ~ obstime + obstime:drug + gender + prevOI + AZT,
      random =  ~ obstime | patient, data = aids)
  jmfit.ew0 <- jm.reg(surv.fit = surv.fit, lme.fit = lme.fit,
      surv.model = "eweibull",  timevar = "obstime", form = "riz")
  newdata <- data.frame(aids[aids$patient == 409, ])
  jm.surv(jmfit.ew0, newdata = newdata, posterior.mean = FALSE)
}
\references{
Rizopoulos D, JMbayes: Joint Modeling of Longitudinal and Time-to-Event Data under a Bayesian Approach, 
          R package version 0.8-85, 2020, \url{https://cran.r-project.org/web/packages/JMbayes}.

Rizopoulos D, The R package JMbayes for fitting joint models for longitudinal and time-to-event data Using MCMC, 
          Journal of Statistical Software, 72(7): 1-45, 2016.
}
\seealso{
\code{\link{jm.lppredict}}, \code{\link{jm.lspredict}}, \code{\link{jm.reg}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Joint}
\keyword{MCMC,}
\keyword{Prediction,}
\keyword{Survival}
\keyword{moldeing,}
\keyword{probabilities}
