% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{LR.test}
\alias{LR.test}
\title{Likelihood ratio test}
\usage{
LR.test(fit)
}
\arguments{
\item{fit}{a list of two \code{etime.reg} fits: one must be the Weibull AFT fit and the other one is the exponentiated
Weibull AFT fit or the generalized gamma AFT fit.}
}
\value{
\code{chi.sq:} chi-square test statistic.

\code{p value:} p value.
}
\description{
Likelihood ratio test for Weibull AFT as a submodel of the exponentiated Weibull AFT and
     generalized gamma AFT models.
}
\details{
The log-likelihood values from the two fits are used for the likelihood ratio test.
}
\examples{
  library(frailtypack)
  data(readmission)
  fit.gg <- etime.reg(c(t.start, t.stop, event) ~ sex + chemo + charlson,
      surv.model = "ggamma", Data = readmission)
  fit.ew <- etime.reg(c(t.start, t.stop, event) ~ sex + chemo + charlson,
      surv.model = "eweibull", Data = readmission)
  fit.w <- etime.reg(c(t.start, t.stop, event) ~ sex + chemo + charlson,
      surv.model = "weibull", Data = readmission)
  LR.test(list(fit.gg, fit.w))
  LR.test(list(fit.ew, fit.w))
}
\seealso{
\code{\link{etime.reg}}, \code{\link{etime.resid}}.
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{Exponentiated}
\keyword{Generalized}
\keyword{Likelihood}
\keyword{Weibull}
\keyword{Weibull,}
\keyword{gamma,}
\keyword{ratio}
\keyword{test,}
