% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMRprint.R
\name{print.JMR}
\alias{print.JMR}
\title{Prints JMR objects}
\usage{
\method{print}{JMR}(x, digits = max(options()$digits - 4, 3))
}
\arguments{
\item{x}{A \code{JMR} object}

\item{digits}{minimal number of significant digits}
}
\value{
No value is returned.
}
\description{
Prints JMR objects
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
