% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrecurrent.R
\name{sim.rec}
\alias{sim.rec}
\title{Simulate recurrent event data}
\usage{
sim.rec(surv.formula, surv.par, surv.model, Data, origin, end.time)
}
\arguments{
\item{surv.formula}{a one-sided formula of the form \code{~ z1 + z2 + ... + zp}, where 
\code{z1, z2, ..., zp} are baseline covariates for the time-to-event process.}

\item{surv.par}{a list of the form \code{list(beta = a vector, logrho = a scalar, logkappa = a scalar)}
for the Weibull AFT, log-logistic AFT, log-normal AFT and Weibull PH models, and
\code{list(beta = a vector, logrho = a scalar, logkappa = a scalar, loggamma = a scalar)}  
for the exponentiated Weibull AFT, generalized gamma AFT and generalized log-logistic PH models,
where \code{beta} is the vector of regression coefficients for the baseline covariates as specified in
\code{surv.formula} (without an intercept), and \code{logrho}, \code{logkappa} and \code{loggamma} are the \code{log} of the
parameters \eqn{\rho}, \eqn{\kappa} and \eqn{\gamma} (see \code{\link{etime.reg}} for 
a description of the distribution parameterizations). Note that \code{surv.par}
must be a named list as shown above.}

\item{surv.model}{the model to be used to describe the recurrent event process. Available options are
\code{"weibull"}, \code{"llogistic"}, \code{"lnormal"}, \code{"eweibull"} and \code{"ggamma"}
for the AFT model, and \code{"weibullph"} and \code{"gllph"} for the PH model.}

\item{Data}{a data frame containing the covariates \code{z1, z2, ..., zp}  for the time-to-event model.
For \eqn{n} subjects, it must must have \eqn{n} rows.}

\item{origin}{the time origin (an \eqn{n} x 1 vector for \eqn{n} subjects; typically, a vector of zeros).}

\item{end.time}{the end of follow-up time (an \eqn{n} x 1 vector for \eqn{n} subjects).}
}
\value{
a data frame containing \code{id}, \code{start}, \code{stop}, \code{status} and covariates.
}
\description{
Simulate recurrent event data for fixed covariates
}
\details{
Simulate recurrent event data. See Section 2.4 of Cook and Lawless.
}
\examples{
  n <- 100 # number of subjects
  Data <- data.frame(z1 = rnorm(n), z2 = rbinom(n, 1, 0.5)) # data frame
  origin <- rep(0, n) # time origin for n subjects
  end.time <- runif(n, 10, 15) # end of follow-up time
  # Simulate data from the generalized gamma AFT model
  surv.par <- list(beta = c(0.5, -0.5), logrho = log(0.025),
      logkappa = log(2), loggamma = log(0.4))
  Data.gg <- sim.rec( ~ z1 + z2, surv.par, "ggamma", Data, origin, end.time)
  # Fit the generalized gamma AFT model
  fit.gg <- etime.reg(c(start, stop, status) ~ z1 + z2,
              surv.model = "ggamma", Data = Data.gg)
  fit.gg
  # Simulate data from the Weibull PH model
  surv.par <- list(beta = c(0.5, -0.5), logrho = log(0.1), logkappa = log(2))
  Data.wph <- sim.rec( ~ z1 + z2, surv.par, "weibullph", Data, origin, end.time)
  # Fit the Weibull PH model
  fit.wph <- etime.reg(c(start, stop, status) ~ z1 + z2,
              surv.model = "weibullph", Data = Data.wph)
  fit.wph
}
\references{
Cook RJ and Lawless J, The statistical analysis of recurrent events, Springer, 2007.
}
\seealso{
\code{\link{etime.reg}}
}
\author{
Shahedul Khan <khan@math.usask.ca>
}
\keyword{AFT}
\keyword{Recurrent}
\keyword{Simulation}
\keyword{data,}
\keyword{event}
\keyword{model,}
